Refactor URL validation classes to add filter
