+++ inc/Engine/Cache/UrlValidation/AbstractUrlValidation.php
@@ -1,0 +1,57 @@
+<?php
+declare(strict_types=1);
+
+namespace WP_Rocket\Engine\Cache\UrlValidation;
+
+abstract class AbstractUrlValidation {
+	/**
+	 * Disable caching invalid page urls.
+	 *
+	 * @param bool $can_cache Filter callback passed value.
+	 *
+	 * @return bool
+	 */
+	public function disable_cache_on_not_valid_pages( $can_cache ) {
+		if ( $this->is_disabled() ) {
+			return $can_cache;
+		}
+
+		if ( $this->is_not_valid_url() ) {
+			return false;
+		}
+
+		return $can_cache;
+	}
+
+	/**
+	 * Stop optimizing those invalid pages by returning empty html string,
+	 * So it fall back to the normal page's HTML.
+	 *
+	 * @param string $html Page's buffer HTML.
+	 *
+	 * @return string
+	 */
+	public function stop_optimizations_for_not_valid_pages( $html ) {
+		if ( $this->is_disabled() ) {
+			return $html;
+		}
+
+		return $this->is_not_valid_url() ? '' : $html;
+	}
+
+	/**
+	 * Check if url validation is disabled by filter
+	 *
+	 * @return bool
+	 */
+	protected function is_disabled(): bool {
+		return wpm_apply_filters_typed( 'boolean', 'rocket_disable_url_validation', false );
+	}
+
+	/**
+	 * Check if current url is not valid
+	 *
+	 * @return bool
+	 */
+	abstract protected function is_not_valid_url(): bool;
+}
